Criação da classe gerenciadora Simulation.java e redistribuição das responsabilidades para isolar a função da Main.
