+++ sicxesimulator/Main.java
@@ -1,0 +1,9 @@
+package sicxesimulator;
+
+import sicxesimulator.simulation.Simulation;
+
+public class Main {
+    public static void main(String[] args) {
+        Simulation simulation = new Simulation();
+    }
+}
